backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
mistralai>=1.0.0
pydantic==2.10.4
aiofiles==24.1.0
psycopg2-binary==2.9.10
